+++ app/controllers/counts_controller.rb
@@ -193,10 +193,17 @@ class CountsController < ApplicationController
   end
   
   def destroy
-    if @count.destroy
-      render json:{status: "success"}, status: 202
+    if current_user.valid_password?(params[:password])
+      if @count.destroy
+        render json:{status: "success"}, status: 202
+      else
+        render json:{status: "error"}, status: 400
+      end
     else
-      render json:{status: "error"}, status: 400
+      render json:{
+        status: "error",
+        message: ["Senha inválida."]
+      }, status: 400
     end
   end
